+++ projects/packages/connection-ui/_inc/selectors/connection-data.js
@@ -1,6 +1,5 @@
 const settingsSelectors = {
 	getDoNotUseConnectionIframe: state => state.connectionData.doNotUseConnectionIframe || null,
-	getAuthorizationUrl: state => state.connectionData.authorizationUrl || null,
 };
 
 export default settingsSelectors;