+++ projects/plugins/jetpack/sal/class.json-api-site-base.php
@@ -150,6 +150,15 @@ abstract class SAL_Site {
 	abstract protected function is_wpforteams_site();
 
 	/**
+	 * Get hub blog id for P2 sites.
+	 *
+	 * @return null
+	 */
+	public function get_p2_hub_blog_id() {
+		return null;
+	}
+
+	/**
 	 * Getter for the p2 organization ID.
 	 *
 	 * @return int
@@ -158,8 +167,13 @@ abstract class SAL_Site {
 		return 0; // WPForTeams\Constants\NO_ORG_ID not loaded.
 	}
 
+	/**
+	 * Detect whether a site is a WordPress.com on Atomic site.
+	 *
+	 * @return bool
+	 */
 	public function is_wpcom_atomic() {
-		return false;
+		return jetpack_is_atomic_site();
 	}
 
 	public function is_wpcom_store() {