+++ projects/js-packages/connection/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ### This is a list detailing changes for the Jetpack RNA Connection Component releases.
 
+## 0.2.0 - 2021-05-25
+### Added
+- Add connection components.
+
+### Changed
+- Updated package dependencies.
+
 ## 0.1.0 - 2021-04-27
 ### Added
-- RNA Connection: InPlaceConnection component added.
+- `InPlaceConnection` component added.
+
+## 0.2.0 - 2021-05-18
+- `Main` and `ConnectUser` components added.
+- `JetpackRestApiClient` API client added.