+++ projects/packages/compat/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.5] - 2021-05-25
+### Changed
+- Updated package dependencies.
+
 ## [1.6.4] - 2021-04-27
 ### Changed
 - Updated package dependencies.
@@ -81,6 +85,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Jetpack 7.5: Back compatibility package
 
+[1.6.5]: https://github.com/Automattic/jetpack-compat/compare/v1.6.4...v1.6.5
 [1.6.4]: https://github.com/Automattic/jetpack-compat/compare/v1.6.3...v1.6.4
 [1.6.3]: https://github.com/Automattic/jetpack-compat/compare/v1.6.2...v1.6.3
 [1.6.2]: https://github.com/Automattic/jetpack-compat/compare/v1.6.1...v1.6.2